requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-bower==5.2.0
django-cors-headers==3.14.0
django-countries==7.5.1
django-htmx==1.14.0
django-paypal==2.0
djangorestframework==3.14.0
frozenlist==1.3.3
geoip2==4.6.0
idna==3.4
jsonpickle==3.0.1
maxminddb==2.2.0
multidict==6.0.4
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python-digitalocean==1.17.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
schedule==1.2.0
six==1.16.0
sqlparse==0.4.3
terraso-allauth==2.0.0
typing-extensions==4.5.0
Unipath==1.1
urllib3==1.26.15
whitenoise==6.4.0
yarl==1.8.2
